requirements.txt

fastapi==0.116.1
uvicorn==0.35.0
pandas
numpy
python-dotenv
